Fix kernel referencing to ensure correct python env is used
This commit adds a line to the _quarto file specifying the jupyter kernel to use. It also creates the _environment.local file specifying the QUARTO_PYTHON path to ensure the correct python is used as well. (that file is placed in the gitignore)

In the future, it would be nice to add the jupyter path to the gitignore as well